Update tag.py
Comparison to singleton `None` should be `expr is None`.

Comparisons to the singleton objects, like True, False, and None, should be done with identity, not equality. Use is or is not.

Identity checks are faster than equality checks.

Also, the equality checks can result in unintended behavior in some cases.